feat(docusaurus-theme): allow controlling global `Demo` scope
